requirements.txt
Django~=5.2.1
djangorestframework~=3.16.0
django-filter~=25.1
drf-yasg~=1.21.10